ewa-urbananalitics-3-developer/UrbanAnalytics/front-end/src/app/dashboard/dashboard.component.spec.ts
import { async, ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';

import { DashboardComponent } from './dashboard.component';
import { RouterTestingModule } from '@angular/router/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { DatasetService } from '../services/dataset/dataset.service';

describe('DashboardComponent', () => {
  let component: DashboardComponent;
  let fixture: ComponentFixture<DashboardComponent>;
  let componentHTML: HTMLElement;
  let datasetService: DatasetService;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [DashboardComponent],
      imports: [
        HttpClientTestingModule,
        RouterTestingModule,
      ]
    })
      .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(DashboardComponent);
    component = fixture.componentInstance;
    componentHTML = fixture.nativeElement;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });


  //  load all the images 
  it('should load all images', () => {
    componentHTML.querySelectorAll('img').forEach(
      (img: HTMLImageElement) => {
        expect(img).toBeTruthy();
        expect(img.complete).toBeTruthy;
      });
  });



});
